devices/Fan/requirements.txt
pyserial>=3.5
requests>=2.30.0
flask>=3.0.0